+++ src/main/java/com/kalsym/locationservice/model/TagTableRequest.java
@@ -16,8 +16,6 @@ public class TagTableRequest {
 
     private String tablePrefix;
 
-    private Integer tableNoStart;
-
-    private Integer tableNoEnd;
+    private String tableNumber;
 
 }